+++ Assets/Scripts/Contents/Note/Note/HoldNoteBody.cs
@@ -24,7 +24,7 @@ public class HoldNoteBody : Note
         DrawLine();
     }
 
-    public override JudgementType ReadJudge(int lane, int bpm, int checkType, float curTime)
+    public override JudgementType ReadJudge(int lane, float bpm, int checkType, float curTime)
     {
         
         if (curJudge == 0)
